Consolidate simiar tests
