+++ EurofurenceTests/Application/Components/Dealers/Presenter Tests/WhenDealersSceneLoads_DealersPresenterShould.swift
@@ -4,7 +4,7 @@ import XCTest
 
 class WhenDealersSceneLoads_DealersPresenterShould: XCTestCase {
 
-    func testBindTheCountOfDealersPerGroupFromTheViewModelOntoTheScene() {
+    func testBindTheDealersAndIndexTitles() {
         let dealerGroups = [DealersGroupViewModel].random
         let viewModel = CapturingDealersViewModel(dealerGroups: dealerGroups)
         let viewModelFactory = FakeDealersViewModelFactory(viewModel: viewModel)
@@ -13,14 +13,6 @@ class WhenDealersSceneLoads_DealersPresenterShould: XCTestCase {
         let expected = dealerGroups.map(\.dealers.count)
 
         XCTAssertEqual(expected, context.scene.capturedDealersPerSectionToBind)
-    }
-
-    func testBindTheSectionIndexTitlesFromTheViewModelOntoTheScene() {
-        let viewModel = CapturingDealersViewModel()
-        let viewModelFactory = FakeDealersViewModelFactory(viewModel: viewModel)
-        let context = DealersPresenterTestBuilder().with(viewModelFactory).build()
-        context.simulateSceneDidLoad()
-
         XCTAssertEqual(viewModel.sectionIndexTitles, context.scene.capturedSectionIndexTitles)
     }
